rename:컴포넌트 관심사에 따라 대략적으로 분리
* /album
   - 앨범
* /review
   - 리뷰
* /common
   - 공통
* /sign
   - 로그인, 회원가입
