+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
-import Header from './components/Header';
+import Header from './components/common/Header';
 import LoginPage from './pages/LoginPage';
 import JoinPage from './pages/JoinPage';
 import DashboardPage from './pages/DashboardPage';
